bytesToWrite now is calculated against a size_t.

+++ modules/c++/sys/source/FileWin32.cpp
@@ -91,15 +91,15 @@ void sys::File::readInto(char *buffer, Size_T size)
 
 void sys::File::writeFrom(const char *buffer, Size_T size)
 {
-    static const DWORD MAX_WRITE_SIZE = std::numeric_limits<DWORD>::max();
+    static const size_t MAX_WRITE_SIZE = std::numeric_limits<DWORD>::max();
     size_t bytesRemaining = size;
     size_t bytesWritten = 0;
 
     while (bytesWritten < size)
     {
         // Determine how many bytes to write
-        const DWORD bytesToWrite =
-            std::min<DWORD>(MAX_WRITE_SIZE, bytesRemaining);
+        const DWORD bytesToWrite = static_cast<DWORD>(
+            std::min<size_t>(MAX_WRITE_SIZE, bytesRemaining));
 
         // Write the data
         DWORD bytesThisWrite = 0;
